commentExceptionAdd

+++ src/main/java/hello/postpractice/service/CommentService.java
@@ -1,5 +1,9 @@
 package hello.postpractice.service;
 
+import hello.postpractice.advice.exception.CommentNotFoundCException;
+import hello.postpractice.advice.exception.PostNotFoundCException;
+import hello.postpractice.advice.exception.UnMatchedUserCException;
+import hello.postpractice.advice.exception.UserNotFoundCException;
 import hello.postpractice.domain.Comment;
 import hello.postpractice.domain.CommentDto;
 import hello.postpractice.domain.Post;
@@ -12,7 +16,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
@@ -25,13 +28,13 @@ public class CommentService {
     @Transactional
     public List<CommentDto> getList(Long postId){
         Post post = postRepository.findById(postId).orElseThrow(() ->
-                new IllegalArgumentException("댓글 쓰기 실패: 해당 게시글이 존재하지 않습니다." + postId));
+                new PostNotFoundCException("댓글 쓰기 실패: 해당 게시글이 존재하지 않습니다." + postId));
         List<CommentDto> commentList = post.getComments().stream().map(CommentDto::new).collect(Collectors.toList());
         return commentList;
     }
     @Transactional
     public CommentDto get(Long commentId){
-        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new IllegalArgumentException("해당 id댓글없음"));
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundCException("해당 id댓글없음"));
         CommentDto commentDto = new CommentDto(comment);
         return commentDto;
     }
@@ -40,9 +43,9 @@ public class CommentService {
     @Transactional
     public CommentDto commentSave(String email, Long postId, CommentDto commentDto) {
         User user = userRepository.findByEmail(email).orElseThrow(()->
-                new IllegalArgumentException("해당 id로 유저 없음"));
+                new UserNotFoundCException());
         Post post = postRepository.findById(postId).orElseThrow(() ->
-                new IllegalArgumentException("댓글 쓰기 실패: 해당 게시글이 존재하지 않습니다." + postId));
+                new PostNotFoundCException("댓글 쓰기 실패: 해당 게시글이 존재하지 않습니다." + postId));
 
         commentDto.setUser(user);
         commentDto.setPost(post);
@@ -54,17 +57,23 @@ public class CommentService {
     }
     /* Update */
     @Transactional
-    public void update(Long commentId, CommentDto commentDto){
-        Comment comment = commentRepository.findById(commentId).orElseThrow(() ->
-                new IllegalArgumentException("해당댓글이 존재하지 않습니다"));
+    public void update(String email, CommentDto commentDto){
+        Comment comment = commentRepository.findById(commentDto.getId()).orElseThrow(() ->
+                new CommentNotFoundCException("해당댓글이 존재하지 않습니다"));
+        if (!(comment.getUser().getEmail() == email)) { //같은 유저 인지 체크
+            throw new UnMatchedUserCException("권한 User가 아닙니다");
+        }
         comment.update(commentDto.getComment());
 
     }
     /* Delete */
     @Transactional
-    public void delete(Long commentId){
+    public void delete(String email,Long commentId){
         Comment comment = commentRepository.findById(commentId).orElseThrow(() ->
-                new IllegalArgumentException("해당 댓글이 존재하지 않습니다"));
+                new CommentNotFoundCException("해당 댓글이 존재하지 않습니다"));
+        if (!(comment.getUser().getEmail() == email)) { //같은 유저 인지 체크
+            throw new UnMatchedUserCException("권한 User가 아닙니다");
+        }
         commentRepository.delete(comment);
     }
 }
